requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
websockets==10.0
pydantic==1.8.2
sqlalchemy==1.4.25
aiosqlite==0.17.0
openai==0.27.0
python-dotenv==0.19.0
httpx==0.23.0
instructor==0.2.1
numpy==1.21.2
sentence-transformers==2.2.0
torch==1.9.0
transformers==4.11.3
rank-bm25==0.2.2
PyPDF2==1.26.0